requirements.txt
# Derin öğrenme
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.0+cu121
torchvision==0.16.0+cu121
torchaudio==2.1.0+cu121

# Görselleştirme ve izleme
tensorboard>=2.16.0
matplotlib>=3.8.0
tqdm>=4.66.0

# Görüntü işleme
opencv-python>=4.9.0
Pillow>=10.0.0
albumentations>=1.4.0

# Bilimsel ve istatistiksel hesaplama
numpy>=1.26.0
scikit-learn>=1.5.0

# Test
pytest>=8.0.0